requirements.txt
streamlit
google-generativeai
pypdf
python-dotenv
requests
sentence-transformers
chromadb
langchain